Resolve "Create an API call, with which to vote on questions"

Closes #39

See merge request cse1105/2020-2021/team-repositories/oopp-group-54/repository-template!33

+++ server/src/main/java/nl/tudelft/oopp/group54/controllers/votes/VoteService.java
@@ -3,8 +3,8 @@ package nl.tudelft.oopp.group54.controllers.votes;
 import java.util.Map;
 
 public interface VoteService {
-    Map<String, Object> voteOnQuestion(Long lectureId,
+    Map<String, Object> voteOnQuestion(Integer lectureId,
                                        String userId,
-                                       Long questionId,
+                                       Integer questionId,
                                        boolean isUpvote);
 }
